+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Bugfixes and Improvements 3.12.0
 
 - TODO: bump AJV and nock
+- CLI: new `mc events` command for event management [#241]
+- CLI: renamed and extended `mc download-events` to `mc events-bulk` command with `--mode download` and  `--mode delete` [#240]
 - Bumped all dependencies
 
 ## 3.11.2 - (Sandcastle Vienna) - March 2021